requirements.txt
pillow>=8.0.0
numpy>=1.20.0
matplotlib>=3.3.0
pandas>=1.2.0
